* src/flucache.c: (slot_available): Fix build in x86_64.
git-svn-id: https://core.fluendo.com/gstreamer/svn/trunk/gst-fluendo-timeshift@2405 2b0047a9-a6d8-0310-accf-f7200b2a168c

+++ src/flucache.c
@@ -84,7 +84,7 @@ struct _Slot
 };
 
 static inline gboolean
-slot_available (Slot * slot, guint * size)
+slot_available (Slot * slot, gsize * size)
 {
   CacheState state = g_atomic_int_get (&slot->state);
